app/requirements.txt
sqlalchemy
mysql-connector-python
streamlit
streamlit-chat
langchain
openai
python-dotenv